streaming_simulation/requirements.txt
# Google Cloud libraries
google-cloud-bigquery==3.11.4
google-cloud-pubsub==2.18.3
google-cloud-storage==2.10.0

# Data processing
pandas==2.0.3
numpy==1.24.3

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Logging and monitoring
colorama==0.4.6